requirements.txt
pandas
scikit-learn
fastapi
uvicorn
joblib
python-multipart
python-dotenv